Store block types and blocks separately.
There's 1 block type for every hash, many (or 0 once frees are implemented)
blocks.

A block type describes behaviour of allocs from that particular hash, stores
hotness data, etc.

A block is a contiguous set of allocations of that type; at this moment
there's a separate block for every alloc, but we'll coalesce them later on
when we start messing with the allocator.